modifiying and moving code to services
