+++ api/mapping/templates/mapping/scanreportvalue_list.html
@@ -82,6 +82,10 @@
             </tbody>
         </table>
 
+	<a href="{% url 'tables-structural-mapping' scan_report.id %}">
+	  <button type="button" class="btn btn-info">Goto Rules</button>
+	</a>
+
     </div>
 
 {% endblock %}
